+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 
+- set users by parts
 - open the conversion dialog box with a gray check mark on the type selection step
 - preload frame moved from uploading popap
 - change tab title on rename