+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+- golang: user role
+- golang: start filling
 - csharp-mvc: user role
 - csharp-mvc: start filling
 - csharp: user role