+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+- ruby: user role
+- ruby: start filling
 - python: user role
 - python: start filling
 - php: user role